requirements.txt
Flask
python-dotenv
Flask-Limiter
Flask-WTF
cloudinary
requests
Pillow
gunicorn